refactor: fix building ID filter in get_building function

+++ Backend/app/models/Building.py
@@ -1,11 +1,10 @@
-# app/models/toilet.py
-from sqlalchemy import Column, Integer, String, Float
+from sqlalchemy import Column, Integer, String, DECIMAL
 from app.db.base import Base
 
 class Building(Base):
     __tablename__ = "building"
     id = Column(Integer, primary_key=True, nullable=False, autoincrement=True)
-    lat = Column(Float, nullable=False)
-    lng = Column(Float, nullable=False)
+    lat = Column(DECIMAL(9, 6), nullable=False)  # 可儲存 ±999.999999
+    lng = Column(DECIMAL(9, 6), nullable=False)
     name = Column(String(20), nullable=False)
     max_floor = Column(Integer, nullable=False)
